requirements.txt
tensorflow-gpu==2.10.0
numpy>=1.19.2
pandas>=1.1.3
scikit-learn>=0.23.2
ta>=0.7.0
joblib>=0.17.0 